refactor(engine): split the fair-share tests at the IOLayer responsibility heading
Addresses the reviewer's post-approval size note (494 > 450 lines): the IOLayer
wrapper tests move to their own module, test_fair_share_io_layer.py, and the
gate fairness-bound test that sat orphaned at the file end now lives under the
fairness-invariants section. Behavior and tests unchanged.

Refs: OME-908
